pyproject.toml
[project]
name = "fastapi-sakms"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.128.2",
    "passlib[argon2]>=1.7.4",
    "pydantic>=2.12.5",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]
